# backend/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Notif API",
        "version": "0.3.0",
        "description": "Personal update monitoring and notification service.\n\nMonitors URLs for changes via pluggable scraping strategies (RSS/Atom feeds, CSS selectors, forum threadmarks, and more)."
    },
    "paths": {
...
        "/api/v1/monitoring/status/": {
            "get": {
                "operationId": "monitoring_status_retrieve",
                "description": "Readiness probe — checks DB connectivity and returns build metadata.\n\nReturns 200 if all dependencies are healthy, 503 otherwise.\nUsed by load balancers and operators to confirm the service can handle traffic\nand to verify which code is deployed.",
                "tags": [
                    "monitoring"
                ],
                "security": [
                    {
                        "sessionBearer": []
                    },
                    {
                        "sessionCookie": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusCheckResponse"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDetail"
                                }
                            }
                        },
                        "description": "The session token is missing where one is required, or is invalid or expired."
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusCheckResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/monitoring/strat-choices": {
            "get": {
                "operationId": "monitoring_strat_choices_list",
                "description": "The strategy class names a Link's Strategy may use, as a flat JSON array. Values match the StratClsEnum used by the Strategy endpoints.",
                "summary": "List available scraping strategy names",
                "tags": [
                    "monitoring"
                ],
                "security": [
                    {
                        "sessionBearer": []
                    },
                    {
                        "sessionCookie": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StratClsEnum"
                                    }
                                }
                            }
                        },
                        "description": "Available strategy class names."
                    },
                    "401": {
                        "description": "Authentication credentials were not provided."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorDetail": {
                "type": "object",
                "description": "DRF's own error body, which every framework-raised 401, 403 and 404 uses.",
                "properties": {
                    "detail": {
                        "type": "string"
                    }
                },
                "required": [
                    "detail"
                ]
            },
...
            "StatusCheckResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "db": {
                        "type": "string"
                    },
                    "version": {
                        "type": "string"
                    },
                    "commit": {
                        "type": "string"
                    },
                    "environment": {
                        "type": "string"
                    }
                },
                "required": [
                    "commit",
                    "db",
                    "environment",
                    "status",
                    "version"
                ]
            },
...
            "StratClsEnum": {
                "enum": [
                    "GeneralSelectorStrategy",
                    "SBSVThreadmarksStrategy",
                    "QQAlertsStrategy",
                    "KemonoFavouritesStrategy",
                    "FeedStrategy"
                ],
                "type": "string",
                "description": "* `GeneralSelectorStrategy` - General Selector\n* `SBSVThreadmarksStrategy` - SB/SV Threadmarks\n* `QQAlertsStrategy` - QQ Alerts\n* `KemonoFavouritesStrategy` - Kemono Favourites\n* `FeedStrategy` - RSS"
            },
...
        },
        "securitySchemes": {
            "sessionBearer": {
                "type": "http",
                "scheme": "session",
                "description": "Opaque session token issued by POST /api/v1/auth/login/ with transport=bearer, sent as `Authorization: Session <token>`. Native clients only."
            },
            "sessionCookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "notif_session",
                "description": "HttpOnly session cookie set by POST /api/v1/auth/login/ with transport=cookie. Browser clients only; unsafe methods must also send the csrftoken value in X-CSRFToken."
            }
        }
    }
}